+++ packages/edit-site/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Code Quality
+
+-   Migrate the Quick Edit panel from `@wordpress/components` `Modal` (with custom drawer-like CSS) to the new `@wordpress/ui` `Drawer` primitive. ([#76837](https://github.com/WordPress/gutenberg/pull/76837))
+
 ## 6.45.0 (2026-04-29)
 
 ## 6.44.0 (2026-04-15)